feat(debug): custom console printing
